requirements.txt
fastapi[standard]
uvicorn
pyjwt
qrcode
python-multipart
cryptography